requirements.txt
pip==21.3.1
pillow==8.4.0